Add the feature stamping run_status when finishing a job

+++ jobs/genome_mapping/aln_3.markdup.sh
@@ -16,16 +16,23 @@ source $(pwd)/$SM/run_info
 set -o nounset
 set -o pipefail
 
+DONE=$SM/run_status/aln_3.markdup.done
+
 printf -- "---\n[$(date)] Start markdup.\n"
 
-$JAVA -Xmx26G -jar $PICARD MarkDuplicates \
-    I=$SM/alignment/$SM.merged.bam \
-    O=$SM/alignment/$SM.markduped.bam \
-    METRICS_FILE=$SM/alignment/markduplicates_metrics.txt \
-    OPTICAL_DUPLICATE_PIXEL_DISTANCE=2500 \
-    CREATE_INDEX=true \
-    TMP_DIR=tmp 
+if [[ -f $DONE ]]; then
+    echo "Skip this step."
+else
+    $JAVA -Xmx26G -jar $PICARD MarkDuplicates \
+        I=$SM/alignment/$SM.merged.bam \
+        O=$SM/alignment/$SM.markduped.bam \
+        METRICS_FILE=$SM/alignment/markduplicates_metrics.txt \
+        OPTICAL_DUPLICATE_PIXEL_DISTANCE=2500 \
+        CREATE_INDEX=true \
+        TMP_DIR=tmp
 
-rm $SM/alignment/$SM.merged.bam{,.bai}
+    rm $SM/alignment/$SM.merged.bam{,.bai}
+    touch $DONE
+fi
 
 printf -- "[$(date)] Finish markdup.\n---\n"
